requirements.txt
# QuidWise Dependencies

# Core
python-dotenv>=1.0.0
pydantic>=2.0.0

# LLM & Agents
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# API Clients
httpx>=0.27.0
yfinance>=0.2.40

# UI
streamlit>=1.38.0
plotly>=5.18.0

# Data Processing
pandas>=2.0.0

# Dev
pytest>=8.0.0
black>=24.0.0
